lagt til service-lag og miljøvariabler
