+++ src/main/resources/application-dev-gcp.yml
@@ -1,5 +1,7 @@
 tiltak-notifikasjon:
   altinn-tilgangstyring:
+    arbtreningServiceCode: 5332
+    arbtreningServiceEdition: 1
     ltsMidlertidigServiceCode: 5516
     ltsMidlertidigServiceEdition: 1
     ltsVarigServiceCode: 5516
@@ -10,5 +12,3 @@ tiltak-notifikasjon:
     mentorServiceEdition: 4
     inkluderingstilskuddServiceCode: 5516
     inkluderingstilskuddServiceEdition: 5
-    arbtreningServiceCode: 5332
-    arbtreningServiceEdition: 1
